update: frontend file structure and navbar buttons
