+++ frontend/app/(landing)/view.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
-import Navbar from '../components/Navbar';
+import Navbar from '../components/navbar/page';
 import { motion } from 'framer-motion';
 import { useRouter } from 'next/navigation';
