+++ frontend/app/components/footer/view.tsx
