+++ frontend/app/page.tsx
@@ -1,9 +1,9 @@
-import Landing from "./landing/page";
+import Landing from "./(landing)/page";
 
 export default function Home() {
   return (
     <div className="main">
-      <Landing />
+      <Landing /> 
     </div>
   )
 }
